Create dynarray_truncate, for setting a dynarray's size.
May be useful for buffer-related operations. Planning for a future
concept of a minimum size hint, to avoid constant growing and shrinking
in case of a buffer.

* dynarray.c (dynarray_truncate): New function.
(dynarray_grow): Renamed to dynarray_append. Now uses dynarray_truncate.
Fixes bug where used_count was being incorrectly set.
* dynarray.h (DYNARRAY_TYPE_DECLARE): Declare dynarray_truncate wrapper.

* include/ncraft.h (enum nc_errno): Documented to be always negative in
case of error. Necessary for functions that overload the return value
e.g. as an index or an error.

* dynarray.c (dynarray_add): Documentation fixes.